attachment_style/requirements.txt
gunicorn
dash-bootstrap-components
dash-mantine-components
dash-iconify
matplotlib
pandas
reportlab
sqlalchemy
alembic
psycopg2
python-dotenv
kaleido
dash_breakpoints